modules/access_control.py
#!/usr/bin/env python3
import boto3
import logging
import os
import subprocess
import re
from concurrent.futures import ThreadPoolExecutor, as_completed
from azure.identity import DefaultAzureCredential
from google.cloud import core
import paramiko
import winrm
import pwd
import grp
from typing import List, Dict, Any, Optional

# Import the new Windows PowerShell scanner
from .windows_powershell import WindowsPowerShellScanner

class AccessControlScanner:
    def __init__(self, config: Dict):
        """
        Initialize Access Control Scanner with advanced security checks
        
        :param config: Configuration dictionary
        """
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Security thresholds and configurations
        self.security_thresholds = {
            'max_privileged_users': 5,
            'max_inactive_days': 90,
            'min_password_complexity': 3,
            'ssh_key_permissions': 0o600,  # Recommended SSH key permissions
            'max_sudo_users': 3
        }

        # Initialize Windows PowerShell scanner
        self.windows_scanner = WindowsPowerShellScanner(config)

    # Existing Linux-specific methods remain the same...

    def _scan_windows_access_controls(self) -> List[Dict]:
        """
        Advanced Windows system access control scanning
        
        :return: List of Windows access control findings
        """
        results = []
        try:
            # Use the new WindowsPowerShellScanner for comprehensive scanning
            results = self.windows_scanner.scan_system_access_controls()

            # Additional custom Windows-specific checks can be added here
            results.append({
                'control_id': 'AC-WINDOWS-CUSTOM',
                'description': 'Custom Windows Access Control Checks',
                'compliant': True,
                'details': {
                    'additional_checks': 'Placeholder for future custom Windows security checks'
                },
                'remediation': 'Continuously update and expand Windows-specific security checks'
            })

        except Exception as e:
            self.logger.error(f"Windows Access Control Scan Error: {e}")
            results.append({
                'control_id': 'AC-WINDOWS-001',
                'description': 'Windows Access Control Scan',
                'compliant': False,
                'remediation': f'Error scanning Windows access controls: {str(e)}'
            })

        return results

    # Rest of the existing methods remain the same...
    # (cloud provider scans, scan method, etc.)
